[iOS] MOVE: 테스트 파일 분리 #208
